Delete unnesecary codes
